Add files via upload
An electric mountain railway ticket system is being developed in Python. Users are able to choose when to go to and from the mountain with this feature. There is a discount available for the tenth person in the group, and consumers are able to choose the hours and number of passengers. If you want to view the total income earned and the most popular train at the conclusion of the day, customers can use the code indefinitely, provided that they register all of their seats. Input, information presentation, and ticket processing are all handled by functions